Sigh

+++ source/http.c
@@ -174,7 +174,7 @@ static void s_init_str_to_enum_hash_table(
         s_destroy_enum_value);
     AWS_FATAL_ASSERT(!err);
 
-    for (size_t i = start_index; i < (size_t)end_index; ++i) {
+    for (int i = start_index; i < end_index; ++i) {
         int was_created = 0;
         struct aws_enum_value *enum_value = aws_mem_calloc(alloc, 1, sizeof(struct aws_enum_value));
         AWS_FATAL_ASSERT(enum_value);
